adding documentation/moving functions to more intuitive place
